feat: 라이더의 기본 기능 완성 - close #80

+++ service/rider/src/main/java/com/webest/rider/presentation/RiderController.java
@@ -1,8 +1,11 @@
 package com.webest.rider.presentation;
 
+import static com.webest.rider.common.constant.StaticVariable.DOCS_AUTH_KEY;
 import static com.webest.web.exception.ErrorCode.INVALID_INPUT;
 
 import com.webest.rider.application.RiderService;
+import com.webest.rider.common.exception.RiderErrorCode;
+import com.webest.rider.common.exception.RiderException;
 import com.webest.rider.domain.model.RiderSearchCondition;
 import com.webest.rider.presentation.dtos.request.RiderCreateRequestDto;
 import com.webest.rider.presentation.dtos.request.RiderRegisterAddressRequestDto;
@@ -11,6 +14,8 @@ import com.webest.rider.presentation.dtos.response.RiderResponseDto;
 import com.webest.web.common.CommonStaticVariable;
 import com.webest.web.common.UserRole;
 import com.webest.web.response.CommonResponse;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -52,12 +57,16 @@ public class RiderController {
     /**
      * 라이더에 속한 주소 등록
      */
-    @PutMapping("/{id}/address")
+    @Operation(
+        summary = "라이더 본인이 원하는 주소 등록",
+        security = @SecurityRequirement(name = DOCS_AUTH_KEY)
+    )
+    @PutMapping("/address")
     public CommonResponse<RiderResponseDto> registerAddressToRider(
-        @PathVariable("id") String riderId,
+        @RequestHeader(name = CommonStaticVariable.X_USER_ID) String userId,
         @RequestBody RiderRegisterAddressRequestDto requestDto
     ) {
-        RiderResponseDto riderResponseDto = riderService.registerAddressToRider(riderId,
+        RiderResponseDto riderResponseDto = riderService.registerAddressToRider(userId,
             requestDto);
 
         return CommonResponse.success(riderResponseDto);
@@ -66,10 +75,20 @@ public class RiderController {
     /**
      * 라이더 단일 조회
      */
-    @GetMapping("/{id}")
+    @Operation(
+        summary = "라이더 단일 조회",
+        security = @SecurityRequirement(name = DOCS_AUTH_KEY)
+    )
+    @GetMapping("/{userId}")
     public CommonResponse<RiderResponseDto> findRiderById(
-        @PathVariable("id") String userId
+        @RequestHeader(name = CommonStaticVariable.X_USER_ID) String xUserId,
+        @RequestHeader(name = CommonStaticVariable.X_USER_ROLE) UserRole userRole,
+        @PathVariable("userId") String userId
     ) {
+        // 자기 자신이나 마스터만 조회 가능
+        if (!userRole.equals(UserRole.MASTER) && !xUserId.equals(userId)) {
+            throw new RiderException(RiderErrorCode.NOT_SELF_USER);
+        }
 
         RiderResponseDto riderById = riderService.findByUserId(userId);
         return CommonResponse.success(riderById);
@@ -78,11 +97,20 @@ public class RiderController {
     /**
      * 라이더 ID로 조회 (마스터만 가능)
      */
+    @Operation(
+        summary = "라이더 ID로 조회",
+        security = @SecurityRequirement(name = DOCS_AUTH_KEY)
+    )
     @GetMapping("/id/{id}")
     public CommonResponse<RiderResponseDto> findRiderById(
         @RequestHeader(name = CommonStaticVariable.X_USER_ROLE) UserRole userRole,
         @PathVariable("id") Long id
     ) {
+
+        // 마스터 확인
+        if (!userRole.equals(UserRole.MASTER)) {
+            throw new RiderException(RiderErrorCode.NOT_ADMIN);
+        }
 
         RiderResponseDto riderById = riderService.findRiderById(id);
         return CommonResponse.success(riderById);
@@ -92,11 +120,20 @@ public class RiderController {
     /**
      * 라이더 검색 (마스터만 가능)
      */
+    @Operation(
+        summary = "라이더 목록 조회",
+        security = @SecurityRequirement(name = DOCS_AUTH_KEY)
+    )
     @GetMapping
     public CommonResponse<Page<RiderResponseDto>> findRiderList(
         @PageableDefault(size = 10, page = 0) Pageable pageable,
-        @ModelAttribute RiderSearchCondition condition
+        @ModelAttribute RiderSearchCondition condition,
+        @RequestHeader(name = CommonStaticVariable.X_USER_ROLE) UserRole userRole
     ) {
+        if (!userRole.equals(UserRole.MASTER)) {
+            throw new RiderException(RiderErrorCode.NOT_ADMIN);
+        }
+
         Page<RiderResponseDto> riderListByCondition = riderService.findRiderListByCondition(
             pageable, condition);
 
@@ -106,9 +143,13 @@ public class RiderController {
     /**
      * 라이디 아이디로 업데이트
      */
-    @PatchMapping("/{id}")
+    @Operation(
+        summary = "라이더 아이디로 업데이트",
+        security = @SecurityRequirement(name = DOCS_AUTH_KEY)
+    )
+    @PatchMapping
     public CommonResponse<?> updateRiderById(
-        @PathVariable("id") String userId,
+        @RequestHeader(name = CommonStaticVariable.X_USER_ID) String userId,
         @Validated @RequestBody RiderUpdateRequestDto requestDto
     ) {
         RiderResponseDto riderResponseDto = riderService.updateRiderById(userId, requestDto);
@@ -119,14 +160,27 @@ public class RiderController {
     /**
      * 라이더 삭제(삭제는 DB ID로 진행한다) ,(마스터만 진행)
      */
+    @Operation(
+        summary = "라이더 삭제(DB ID로 삭제를 진행한다.",
+        security = @SecurityRequirement(name = DOCS_AUTH_KEY)
+    )
     @DeleteMapping("/{id}")
     public CommonResponse<Long> deleteRiderById(
+        @RequestHeader(name = CommonStaticVariable.X_USER_ROLE) UserRole userRole,
         @PathVariable("id") Long id
     ) {
+        if (!userRole.equals(UserRole.MASTER)) {
+            throw new RiderException(RiderErrorCode.NOT_ADMIN);
+        }
+
         Long deletedId = riderService.deleteRiderById(id);
 
         return CommonResponse.success(deletedId);
     }
+
+    /* ================================================================================
+      Private Method
+     ================================================================================*/
 
     /* ================================================================================
       Exception Handler
